Manually modified kernel for AC: new version to make comparison fair.
Our previous version contains allocators for int and pointer, which is
unnecessary for this kernel. In order to make our comparision fair, we
remove these code and simply use fix size arrays. We also simplified a
recursion conversion because it can be actually convert to a tail
recursion then a while block. The code length of this file is therefore
significantly different from the one reported in the draft.